refactor: reorganize egghead api utilities into a modular structure
Breaking down egghead.ts into smaller, focused modules for better maintainability.
- Created type definitions file for schemas and interfaces
- Separated authentication, resources, lessons, courses, tags, and utilities
- Fixed async typing issues in write.ts and utility functions
- Created index.ts for re-exports and backward compatibility

Improves code organization and makes future changes easier to manage.
