+++ apps/egghead/src/lib/egghead/index.ts
@@ -1,0 +1,24 @@
+/**
+ * Egghead API integration module - handles interactions with Egghead.io platform
+ */
+
+// Re-export types
+export * from './types'
+
+// Re-export authentication-related functions
+export * from './auth'
+
+// Re-export resource-related functions
+export * from './resource'
+
+// Re-export lesson-related functions
+export * from './lesson'
+
+// Re-export course-related functions
+export * from './course'
+
+// Re-export tagging-related functions
+export * from './tags'
+
+// Re-export utility functions
+export * from './utils'
